+++ source/jpegdata/enumerations/__init__.py
@@ -2,8 +2,10 @@ from jpegdata.enumerations.format import Format
 from jpegdata.enumerations.encoding import Encoding
 from jpegdata.enumerations.marker import Marker
 from jpegdata.enumerations.state import State
+from jpegdata.enumerations.transform import ColourTransform
 
 __all__ = [
+    "ColourTransform",
     "Format",
     "Encoding",
     "Marker",